+++ FYP/Xcode/SmartHawker/SmartHawker/AnalyticsViewController.swift
@@ -207,7 +207,7 @@ class AnalyticsViewController: UIViewController, ChartViewDelegate {
                     for object in objects {
                         let dateString = object["date"] as! String
                         let type = object["type"] as! Int
-                        let amount = object["amount"] as! Int
+                        let amount = object["amount"] as! Double
                         var localIdentifierString = object["subUser"]
                         var typeString = ""
                         if (type == 0) {
