+++ FYP/Xcode/SmartHawker/SmartHawker/MainViewController.swift
@@ -256,7 +256,7 @@ class MainViewcontroller: UIViewController, CLLocationManagerDelegate{
                     for object in objects {
                         let dateString = object["date"] as! String
                         let type = object["type"] as! Int
-                        let amount = object["amount"] as! Int
+                        let amount = object["amount"] as! Double
                         var localIdentifierString = object["subUser"]
                         var typeString = ""
                         if (type == 0) {
