+++ FYP/Xcode/SmartHawker/SmartHawker/RecordSalesViewController.swift
@@ -87,7 +87,7 @@ class RecordSalesViewController: UIViewController, UITextFieldDelegate {
     
     func SubmitRecord(completionHandler: CompletionHandler) {
         let descriptionToRecord = descriptionTextField.text
-        let amountToRecord = Int(amountTextField.text!)
+        let amountToRecord = Double(amountTextField.text!)
         var didRecord = false
         let isSubUser = shared.isSubUser
